[create] - criação da camada de validação e regras de negócios.
