+++ src/main/java/br/com/alura/comunidade/domain/dto/TopicoDtoDetalhar.java
