+++ src/main/java/br/com/alura/comunidade/domain/model/StatusTopico.java
@@ -1,0 +1,15 @@
+/**
+ * User: phmiranda
+ * Project: comunidade
+ * Description: this class execute...!
+ * Date: 23/12/2020
+ */
+
+package br.com.alura.comunidade.entity.model;
+
+public enum StatusTopico {
+	NAO_RESPONDIDO,
+	NAO_SOLUCIONADO,
+	SOLUCIONADO,
+	FECHADO;
+}
