+++ src/main/java/br/com/alura/comunidade/service/TopicoService.java
@@ -1,0 +1,12 @@
+/**
+ * User: Pedro
+ * Project: comunidade
+ * Description: this class execute...!
+ * Date: 29/04/2021
+ */
+
+package br.com.alura.comunidade.service;
+
+public class TopicoService {
+    // comment
+}
